Asynchronous node bindings that work
This commit introduces asynchronous node bindings that work [*]. It has been
tested on Linux and macOS with MK v0.9.0-dev and v0.7.10.

It has issues with MK v0.7.10. Specifically, the main problem seems to be
that tests don't terminate properly closing the libuv main loop.

This is based on #2 started by @hellais. I squashed all the commits
together. For the full history, see #2.

[*] Exceptions may exist! Consult a laywer! If in doubt, don't breathe!