Rewrite async context to be a namespace
It has been pointed out by @hellais and @pboothe that using a class
with static methods only is worst than having a namespace that groups
all the related functionality, so rewrite the code for clarity.

A follow-up commit will also make more intuitive the way in which
we keep alive the object on the C side. This would probably reduce
one additional place where we do confusing things.

Started doing this here because this code is relatively small, yet it
has similar problems to MK, so doing changes here helps me to understand
how porting specific patterns would like like in MK.