fix(async): less confusing patter for keeping-it-alive
Rather than having a self-reference inside the structure itself, allocate
a shared pointer on the heap and pass it to C code as `void *` field.

This is much more consistent with the way in which stuff is typically passed
to C code and hence it should produce the least surprise.